refactor(SnackBar): refactor all css files and component structure
closes #1821
