--- src/components/SnackBar/ActionButton/SnackBar-ActionButton.css
@@ -1,11 +1,0 @@
-.SnackBar {
-  &-ActionButtonsWrapper {
-    padding-top: var(--space-2xs);
-  }
-
-  &-ActionButton {
-    margin-top: var(--space-xs);
-    margin-right: var(--space-xs);
-    filter: drop-shadow(0 0 0 var(--color-bg-soft));
-  }
-}
