--- src/components/SnackBar/Item/SnackBar-Item.css
@@ -1,92 +1,0 @@
-.SnackBar {
-  &-Item {
-    position: relative;
-    display: flex;
-    margin-bottom: var(--space-s);
-    padding: var(--space-m) var(--space-2xl) var(--space-m) var(--space-m);
-    color: var(--color-typo-primary);
-    border-radius: var(--space-xs);
-    box-shadow: var(--shadow-group);
-
-    &_showProgress {
-      &_progress-line {
-        padding-bottom: calc(var(--space-m) + var(--space-2xs));
-      }
-    }
-
-    &_status {
-      &_system {
-        background-color: var(--color-bg-system);
-      }
-
-      &_normal {
-        background-color: var(--color-bg-normal);
-      }
-
-      &_success {
-        background-color: var(--color-bg-success);
-      }
-
-      &_warning {
-        background-color: var(--color-bg-warning);
-      }
-
-      &_alert {
-        background-color: var(--color-bg-alert);
-      }
-    }
-  }
-
-  &-Icon {
-    flex: none;
-    margin-right: var(--space-m);
-  }
-
-  &-Content {
-    flex: 1;
-  }
-
-  &-CloseButton {
-    position: absolute;
-    top: var(--space-2xs);
-    right: var(--space-2xs);
-  }
-
-  &-ProgressLine {
-    position: absolute;
-    left: 0%;
-    bottom: 0;
-    overflow: hidden;
-    width: 100%;
-    height: var(--space-m);
-    border-radius: var(--space-xs);
-    opacity: 0.8;
-
-    &::after {
-      content: '';
-      position: absolute;
-      left: 0%;
-      bottom: 0;
-      width: 200%;
-      height: var(--space-2xs);
-      background: linear-gradient(
-        90deg,
-        var(--color-bg-default) 0%,
-        var(--color-bg-default) 50%,
-        var(--color-bg-ghost) 50%,
-        var(--color-bg-ghost) 100%
-      );
-      animation: Progress-Line var(--progress-line-time) linear;
-    }
-  }
-}
-
-@keyframes Progress-Line {
-  0% {
-    left: -100%;
-  }
-
-  100% {
-    left: 0%;
-  }
-}
