+++ src/components/SnackBar/SnackBar.css
@@ -6,39 +6,4 @@
   max-width: calc(400px + var(--space-xs) * 2);
   padding-right: var(--space-xs);
   padding-left: var(--space-xs);
-
-  &-Item {
-    &_animate {
-      &_appear,
-      &_enter,
-      &_exit,
-      &_appearActive,
-      &_enterActive,
-      &_exitActive {
-        transition: transform 0.2s, opacity 0.2s;
-      }
-
-      &_appear,
-      &_enter {
-        opacity: 0;
-        transform: scale(0.5);
-      }
-
-      &_appearActive,
-      &_enterActive {
-        opacity: 1;
-        transform: scale(1);
-      }
-
-      &_exit {
-        opacity: 1;
-        transform: scale(1);
-      }
-
-      &_exitActive {
-        opacity: 0;
-        transform: scale(0.5);
-      }
-    }
-  }
 }
