+++ src/components/SnackBar/SnackBarActionButton/SnackBarActionButton.css
@@ -1,0 +1,9 @@
+.SnackBarActionButton {
+  margin-top: var(--space-xs);
+  margin-right: var(--space-xs);
+  filter: drop-shadow(0 0 0 var(--color-bg-soft));
+
+  &-ActionButtonsWrapper {
+    padding-top: var(--space-2xs);
+  }
+}
