+++ src/components/SnackBar/SnackBarItem/SnackBarItem.css
@@ -1,0 +1,85 @@
+.SnackBarItem {
+  position: relative;
+  display: flex;
+  margin-bottom: var(--space-s);
+  padding: var(--space-m) var(--space-2xl) var(--space-m) var(--space-m);
+  color: var(--color-typo-primary);
+  border-radius: var(--space-xs);
+  box-shadow: var(--shadow-group);
+
+  &_showProgress {
+    &_line {
+      padding-bottom: calc(var(--space-m) + var(--space-2xs));
+    }
+  }
+
+  &_status {
+    &_system {
+      background-color: var(--color-bg-system);
+    }
+
+    &_normal {
+      background-color: var(--color-bg-normal);
+    }
+
+    &_success {
+      background-color: var(--color-bg-success);
+    }
+
+    &_warning {
+      background-color: var(--color-bg-warning);
+    }
+
+    &_alert {
+      background-color: var(--color-bg-alert);
+    }
+  }
+
+  &_animate {
+    &_appear,
+    &_enter,
+    &_exit,
+    &_appearActive,
+    &_enterActive,
+    &_exitActive {
+      transition: transform 0.2s, opacity 0.2s;
+    }
+
+    &_appear,
+    &_enter {
+      opacity: 0;
+      transform: scale(0.5);
+    }
+
+    &_appearActive,
+    &_enterActive {
+      opacity: 1;
+      transform: scale(1);
+    }
+
+    &_exit {
+      opacity: 1;
+      transform: scale(1);
+    }
+
+    &_exitActive {
+      opacity: 0;
+      transform: scale(0.5);
+    }
+  }
+
+  &-Icon {
+    flex: none;
+    margin-right: var(--space-m);
+  }
+
+  &-Content {
+    flex: 1;
+  }
+
+  &-CloseButton {
+    position: absolute;
+    top: var(--space-2xs);
+    right: var(--space-2xs);
+  }
+}
