+++ src/components/SnackBar/SnackBarTimer/SnackBarTimer.css
@@ -1,0 +1,4 @@
+.SnackBarTimer {
+  flex: none;
+  margin-right: var(--space-m);
+}
