+++ src/components/SnackBar/__stories__/examples/SnackBarExampleButtons/SnackBarExampleButtons.tsx
@@ -4,7 +4,8 @@ import React from 'react';
 
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
 import { cn } from '../../../../../utils/bem';
-import { Item, SnackBar } from '../../../SnackBar';
+import { Item } from '../../../helper';
+import { SnackBar } from '../../../SnackBar';
 
 const cnSnackBarExampleButtons = cn('SnackBarExampleButtons');
