+++ src/components/SnackBar/__stories__/examples/SnackBarExampleIcon/SnackBarExampleIcon.tsx
@@ -6,7 +6,8 @@ import { IconDisconnection } from '../../../../../icons/IconDisconnection/IconDi
 import { IconMoon } from '../../../../../icons/IconMoon/IconMoon';
 import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
 import { cn } from '../../../../../utils/bem';
-import { Item, SnackBar } from '../../../SnackBar';
+import { Item } from '../../../helper';
+import { SnackBar } from '../../../SnackBar';
 
 const cnSnackBarExampleIcon = cn('SnackBarExampleIcon');
