+++ src/uiKit/components/ThemePreview/ThemePreview.tsx
@@ -2,7 +2,8 @@ import './ThemePreview.css';
 
 import React, { useReducer } from 'react';
 
-import { Item, SnackBar } from '../../../components/SnackBar/SnackBar';
+import { Item } from '../../../components/SnackBar/helper';
+import { SnackBar } from '../../../components/SnackBar/SnackBar';
 import { Text } from '../../../components/Text/Text';
 import { defaultVars } from '../../../hooks/useThemeVars/helpers';
 import { useThemeVars } from '../../../hooks/useThemeVars/useThemeVars';
